Setting the preconfigured log level to critical. The documentation directs users to reset log leves to DEBUG if they require it.
Setting the log level to CRITICAL will not log any messages from
DEBUG through to ERROR, hence avoiding the overhead of priting
debud messages as defined in the modeule.

Users can set the log level to DEBUG if they wish to see debug
messages and reset the log level to CRITICAL in production so as
to avoid logging.